requirements.txt
customtkinter==5.2
opencv-contrib-python
supabase
python-dotenv
Pillow>=10.2.0
numpy
